[#216] ✨ Integrate API headers for login, logout, and registration endpoints
